+++ include/ulog.h
@@ -303,6 +303,12 @@ ulog_status ulog_output_remove(ulog_output_id output);
 /// @return Topic ID on success, ULOG_TOPIC_ID_INVALID on failure
 ulog_topic_id ulog_topic_add(const char *topic_name, ulog_output_id output,
                              bool enable);
+                             
+/// @brief Removes a topic  (requires ULOG_BUILD_TOPICS!=0 or
+/// ULOG_BUILD_DYNAMIC_CONFIG=1)
+/// @param topic_name Topic name string (empty or NULL names are invalid)
+/// @return ULOG_STATUS_OK on success, ULOG_STATUS_NOT_FOUND if topic not found
+ulog_status ulog_topic_remove(const char *topic_name);
 
 /// @brief Sets the minimum log level for a topic  (requires
 /// ULOG_BUILD_TOPICS!=0 or ULOG_BUILD_DYNAMIC_CONFIG=1)